Add `mingus-get-songs-with-smart-cache' and use it with ido.
This differs from `mingus-get-songs' in that it only requests MPD
for the information when the database has another version number.

If versions differ, the cached response is used.

NOTE that the playlist may differ while the database may not, so
do not use this function for stuff concerning the playlist.

This function may fail on the off chance servers are switched and
their latest update time happen to be exactly the same. In that
rare case, running a single `mingus-update' to regenerate the
database may work.

Currently ido-mode completion will use this.