Hopefully fix some redisplay/lag issues.
This avoids populating our initial cache for searches, to avoid the
lag when starting mingus. Instead, it will just create cache on
demand, i.e. when someone performs a search for the first time.

It also puts some stuff in the modeline object instead of the timer.
This means the modeline timer does not get updated as frequently in
non-playlist buffers (more precisely: it happens only on buffer
changes/user interaction), but flickering seems to be gone.