* mingus.el: Simplify code adding streams in podcast, and allow querystrings.
* mingus.el (mingus-add-podcast): use helper function
`mingus-streams-from-podcast-with-regexp' to retrieve all podcasts
from the podcast xml string.

Also include for possible future use a commented out example using
xml-parse and s-query (which would have to be edited to become better
suited to output from xml-parse, increase dependencies, and of which I
haven't inquired into its copyright and licences ... but would of
course provide a more robust solution).