Fallback to locally saving playlists.
Mopidy currently does not support saving m3u files through the MPD
interface. So I thought it easiest to just create an m3u file ourselves.

NOTE that mopidy currently needs to be restarted to load the m3u file.

NOTE also that when `loading' the m3u file no id3 info gets used in the
playlist, but when first `listing' the playlist and then selecting the
separate files (e.g. by selecting the entire buffer) and loading them,
this id3 information *is* represented in the playlist.