Bump version to celebrate the mopidy branch merge
This version has the nickname Alice's Wonderland.

From the liner notes on the back of Wonderland:

"This", says Mingus, "is a portrait of a girl trying to make it in this
big, rough world, like I am. I try to show her sadness (the alto on top)
but also her strength in her art and in her conviction in what she
believes in (the tenor on the bottom) even if there are still harsh,
unresolved parts of her life."

Something like that might very well be applicable to this release. Stuff
may turn out rough for old-time users*, but I do think this is the way
forward.

Let be know if there are any 'unresolved parts'.

And listen to the song.

P.S. Check out https://gist.github.com/pft/684603b6cc2856e41ac1 if you
want automatic resizing of Mingus buffers.

*There have been some code clean-ups, changes of text-properties (now
 the same in Playlist and in Browser), and a column layout has
 supplanted old formatting functions.