Make current song props fit every color scheme
You can now customize `mingus-current-song-props' in order to optionally
set your own background color or don't use a bold face.

Thanks to mgi and the pull request at
https://github.com/pft/mingus/pull/24/commits/77cb8da49fa7593ea02566ab1dcb48567dac6410
for pointing out incompatibility of the previous default darkslategray
with some color schemes.

If anyone understands how to create a better looking customization
widget for this I am open to it.